Update README.md
About me:
I’m talented and dependable video-maker skilled in all aspects of video-making and psychology of attention. 
I know how to keep the audience hooked on the video they are watching. 
I’m Highly organized and detail-oriented. An expert in Premiere Pro, High Level in After Effects and Basic Level in Photoshop with over 5 years of experience. 
I have strong background in modern day media, keep up-to-date with changes in the industry and have an ability to find innovative solutions to any kind of situation.